refactor dependency table
