+++ src/components/DependencyChart/DependencyChart.js
@@ -1,18 +1,14 @@
 import React, { useContext } from 'react';
 import DependencyTable from './DependencyTable';
 import AppsContext from '../../contexts/AppsContext';
-import { Link } from 'react-router-dom';
+import { createAppList, createDependencyList } from '../../helpers/chartdata';
 import '../../styles/DependencyChart.scss';
 
 function DependencyChart() {
   const { apps } = useContext(AppsContext);
-  const appList = [];
-  for (let i = 0; i < apps.length; i++) {
-    appList.push({
-      Header: <Link to={`/app/${apps[i].id}`}>{apps[i].id}</Link>,
-      accessor: `${apps[i].id}`,
-    });
-  }
+
+  const appList = createAppList(apps);
+  const depList = createDependencyList(apps, appList);
 
   const columns = [
     {
@@ -30,35 +26,6 @@ function DependencyChart() {
     },
   ];
 
-  const depList = [];
-  const deps = [];
-  for (let i = 0; i < apps.length; i++) {
-    let result = apps[i].dependencies;
-    for (let j = 0; j < result.length; j++) {
-      if (!deps.includes(result[j].name)) {
-        deps.push(result[j].name);
-      }
-    }
-  }
-
-  for (let i = 0; i < deps.length; i++) {
-    depList.push({
-      name: deps[i],
-    });
-  }
-
-  for (let i = 0; i < appList.length; i++) {
-    let index = 0;
-    for (let j = 0; j < depList.length; j++) {
-      if (!apps[i].dependencies[index]) depList[j][apps[i].id] = 'N/A';
-      else if (depList[j].name === apps[i].dependencies[index].name) {
-        depList[j][apps[i].id] = apps[i].dependencies[index].version;
-        index++;
-      } else {
-        depList[j][apps[i].id] = 'N/A';
-      }
-    }
-  }
   if (apps.length && columns.length)
     return (
       <div className="DependencyChart">
